% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{add_batting_derived_metrics}
\alias{add_batting_derived_metrics}
\title{Add derived batting metrics.}
\usage{
add_batting_derived_metrics(df)
}
\arguments{
\item{df}{Normalized batting player tibble.}
}
\value{
Input tibble with derived batting and value metrics appended.
}
\description{
Add derived batting metrics.
}
