% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{add_pitching_derived_metrics}
\alias{add_pitching_derived_metrics}
\title{Add derived pitching metrics.}
\usage{
add_pitching_derived_metrics(df)
}
\arguments{
\item{df}{Normalized pitching player tibble.}
}
\value{
Input tibble with derived pitching and value metrics appended.
}
\description{
Add derived pitching metrics.
}
