% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_engine.R
\name{ncaa_aggregate_qualified_team_batting}
\alias{ncaa_aggregate_qualified_team_batting}
\title{Aggregate qualified team-level batting stats.}
\usage{
ncaa_aggregate_qualified_team_batting(df, min_pa = 15)
}
\arguments{
\item{df}{Batting player-level data frame.}

\item{min_pa}{Minimum plate appearances to qualify a hitter.}
}
\value{
Team-level qualified batting tibble.
}
\description{
Aggregate qualified team-level batting stats.
}
