% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_engine.R
\name{ncaa_aggregate_qualified_team_pitching}
\alias{ncaa_aggregate_qualified_team_pitching}
\title{Aggregate qualified team-level pitching stats.}
\usage{
ncaa_aggregate_qualified_team_pitching(df, min_ip = 5)
}
\arguments{
\item{df}{Pitching player-level data frame.}

\item{min_ip}{Minimum innings pitched to qualify a pitcher.}
}
\value{
Team-level qualified pitching tibble.
}
\description{
Aggregate qualified team-level pitching stats.
}
