% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_engine.R
\name{ncaa_aggregate_team_batting}
\alias{ncaa_aggregate_team_batting}
\title{Aggregate team-level batting stats from player rows.}
\usage{
ncaa_aggregate_team_batting(df)
}
\arguments{
\item{df}{Batting player-level data frame.}
}
\value{
Team-level batting tibble.
}
\description{
Aggregate team-level batting stats from player rows.
}
