% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ncaa_default_config}
\alias{ncaa_default_config}
\title{Default sync configuration.}
\usage{
ncaa_default_config()
}
\value{
Named list used by \code{ncaa_sync_daily()}.
}
\description{
Default sync configuration.
}
